@@ -1,3 +1,3 @@
-import {ModI18N} from "./I18N";
+import { ModI18N } from "./I18N";
 
 window.modI18N = new ModI18N();
