@@ -1,13 +1,13 @@
-import type {SC2DataManager} from "../../../dist-BeforeSC2/SC2DataManager";
-import type {ModUtils} from "../../../dist-BeforeSC2/Utils";
+import type { SC2DataManager } from "../lib/sugarcube-2-ModLoader/dist-BeforeSC2/SC2DataManager";
+import type { ModUtils } from "../lib/sugarcube-2-ModLoader/dist-BeforeSC2/Utils";
 
 declare global {
-    interface Window {
-        modUtils: ModUtils;
-        modSC2DataManager: SC2DataManager;
+  interface Window {
+    modUtils: ModUtils;
+    modSC2DataManager: SC2DataManager;
 
-        modI18N: ModI18N;
-    }
+    modI18N: ModI18N;
+  }
 
-    var StartConfig: any;
+  var StartConfig: any;
 }
